refactor(store): integrate store client module (#47)
* refactor(store): Added query pushdown support for Server & PD

- Add StreamObserver implementation for server side
- Modified the visibility of member variables to support query pushdown
- Add HgSessionConfig.java

* refactor(store): Support get partition from pd based on graph name & code & start key

* chore(store): reformat code & code cleanup

* fix(store): fix problems in code review

fix(store): fix unused sessionConfig in HgStoreClient.java

fix(store): fix potential NPE exception

fix(store): Fix incorrect spelling

fix(store): Fix the unit inconsistency in the time comparison

Update hugegraph-store/hg-store-client/src/main/java/org/apache/hugegraph/store/client/query/QueryExecutor.java


Update hugegraph-store/hg-store-client/src/main/java/org/apache/hugegraph/store/client/query/StreamSortedIterator.java


Update hugegraph-store/hg-store-client/src/main/java/org/apache/hugegraph/store/client/query/MultiStreamIterator.java


Update hugegraph-store/hg-store-client/src/main/java/org/apache/hugegraph/store/client/query/StreamFinalAggregationIterator.java


* chore(store): remove unused code & modify import path

* fix(store): fix bug in ut

* fix(store): add missing table statement

* fix(store): fix ci problem

* Trigger ci

---------

+++ hugegraph-store/hg-store-client/src/main/java/org/apache/hugegraph/store/client/SequencedIterator.java
@@ -51,7 +51,7 @@ public class SequencedIterator implements HgKvIterator {
     SequencedIterator(List<HgKvOrderedIterator> iterators, long limit) {
         Collections.sort(iterators);
         this.queue = new LinkedList(iterators);
-        this.limit = limit <= 0 ? Integer.MAX_VALUE : limit;
+        this.limit = limit <= 0 ? Long.MAX_VALUE : limit;
     }
 
     private HgKvOrderedIterator getIterator() {
